backend/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.3
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg>=0.29.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.20
python-dotenv==1.0.0
requests>=2.32.3
httpx>=0.25.0
tenacity>=8.2.3
itsdangerous==2.1.2
fastapi-mail==1.4.1
line-bot-sdk==3.13.0
redis>=5.0.1
cachetools>=5.3.0
aiohttp>=3.9.0
websockets>=12.0
minio>=7.2.0
aiofiles>=24.1.0
motor==3.3.2
pymongo==4.6.1
beanie==1.23.6
groq>=0.4.1
pgvector>=0.2.5
sentence-transformers>=2.2.2
pdfplumber>=0.11.4
PyPDF2>=3.0.1
python-docx>=1.1.0
